Move sheets into independent library
