+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,6 @@
         <activity android:name=".Spiral" />
         <activity android:name=".Level" />
         <activity android:name=".Popper" />
-        <activity android:name=".Sheets"></activity>
         <activity android:name=".Curling"></activity>
     </application>
