+++ app/src/main/java/com/example/tapp/Spiral.java
@@ -19,6 +19,8 @@ import android.view.View;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.example.sheets436.Sheets;
+
 import java.util.Timer;
 import java.util.TimerTask;
