+++ sheets436/src/main/res/layout/activity_sheets.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.tapp.Sheets">
+    tools:context="com.example.sheets436.Sheets">
 
 </android.support.constraint.ConstraintLayout>
